requirements-dev.txt
build
twine
coverage
mkdocs-material
mkdocstrings
mkdocstrings-python
pytest
pytest-django
pytest-cov
